mithril_cardano_node_chain

Module chain_importer

Tools to import Cardano chain data (Cardano transactions, Block range with their computed merkle root) into a data store.

Structs

CardanoChainDataImporter

Import and store Cardano chain data (transactions, blocks) and aggregate them into Merkle trees for Mithril certificate generation and proof verification.

ChainDataImporterByChunk

A decorator of ChainDataImporter that does the import by chunks

ChainDataImporterWithPruner

A decorator of ChainDataImporter that prunes the chain data older than a given number of blocks after running the import.

Traits

ChainDataImporter

Cardano chain data importer

ChainDataPruner

Cardano stored chain data pruner

ChainDataStore

Cardano chain data store

HighestStoredBlockNumberGetter

Trait to get the highest stored chain data block number