mithril_cardano_node_chain

Module chain_observer

Tools to request metadata, like the current epoch or the stake distribution, from the Cardano

Structs

CardanoCliChainObserver

A ChainObserver pulling it’s data using a CardanoCliRunner.

CardanoCliRunner

A runner able to request data from a Cardano node using the Cardano Cli.

ChainObserverBuilder

Chain observer builder

PallasChainObserver

A runner that uses Pallas library to interact with a Cardano node using N2C Ouroboros mini-protocols

Enums

ChainObserverBuilderError

Error type for chain observer builder service.

ChainObserverError

ChainObserver related errors.

ChainObserverType

Type of chain observers available

Traits

ChainObserver

Retrieve data from the cardano network

CliRunner

CliRunner trait defines the asynchronous methods for interaction with the Cardano CLI.