mithril_cardano_node_chain::chain_importer

Trait ChainDataStore

pub trait ChainDataStore: Send + Sync {
    // Required methods
    fn get_highest_beacon<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = StdResult<Option<ChainPoint>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_highest_block_range<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = StdResult<Option<BlockRange>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_highest_legacy_block_range<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = StdResult<Option<BlockRange>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn store_blocks_and_transactions<'life0, 'async_trait>(
        &'life0 self,
        block_with_transactions: Vec<CardanoBlockWithTransactions>,
    ) -> Pin<Box<dyn Future<Output = StdResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_blocks_and_transactions_in_range<'life0, 'async_trait>(
        &'life0 self,
        range: Range<BlockNumber>,
    ) -> Pin<Box<dyn Future<Output = StdResult<BTreeSet<CardanoBlockTransactionMkTreeNode>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_transactions_in_range<'life0, 'async_trait>(
        &'life0 self,
        range: Range<BlockNumber>,
    ) -> Pin<Box<dyn Future<Output = StdResult<Vec<CardanoTransaction>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn store_block_range_roots<'life0, 'async_trait>(
        &'life0 self,
        block_ranges: Vec<(BlockRange, MKTreeNode)>,
    ) -> Pin<Box<dyn Future<Output = StdResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn store_legacy_block_range_roots<'life0, 'async_trait>(
        &'life0 self,
        block_ranges: Vec<(BlockRange, MKTreeNode)>,
    ) -> Pin<Box<dyn Future<Output = StdResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn remove_rolled_chain_data_and_block_range<'life0, 'async_trait>(
        &'life0 self,
        slot_number: SlotNumber,
    ) -> Pin<Box<dyn Future<Output = StdResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn optimize<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = StdResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Cardano chain data store

Required Methods

fn get_highest_beacon<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = StdResult<Option<ChainPoint>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the highest known transaction beacon

fn get_highest_block_range<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = StdResult<Option<BlockRange>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the highest stored block range root bounds

fn get_highest_legacy_block_range<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = StdResult<Option<BlockRange>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the highest stored legacy block range root bounds

fn store_blocks_and_transactions<'life0, 'async_trait>( &'life0 self, block_with_transactions: Vec<CardanoBlockWithTransactions>, ) -> Pin<Box<dyn Future<Output = StdResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Store the given blocks and their transactions

fn get_blocks_and_transactions_in_range<'life0, 'async_trait>( &'life0 self, range: Range<BlockNumber>, ) -> Pin<Box<dyn Future<Output = StdResult<BTreeSet<CardanoBlockTransactionMkTreeNode>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get all blocks and transactions in an interval of blocks

fn get_transactions_in_range<'life0, 'async_trait>( &'life0 self, range: Range<BlockNumber>, ) -> Pin<Box<dyn Future<Output = StdResult<Vec<CardanoTransaction>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get transactions in an interval of blocks

fn store_block_range_roots<'life0, 'async_trait>( &'life0 self, block_ranges: Vec<(BlockRange, MKTreeNode)>, ) -> Pin<Box<dyn Future<Output = StdResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Store list of block ranges with their corresponding merkle root

fn store_legacy_block_range_roots<'life0, 'async_trait>( &'life0 self, block_ranges: Vec<(BlockRange, MKTreeNode)>, ) -> Pin<Box<dyn Future<Output = StdResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Store list of legacy block ranges with their corresponding merkle root

fn remove_rolled_chain_data_and_block_range<'life0, 'async_trait>( &'life0 self, slot_number: SlotNumber, ) -> Pin<Box<dyn Future<Output = StdResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Remove blocks, transactions, and block range roots that are in a rolled-back fork

Remove blocks and transactions with a slot number strictly greater than the given slot number

Remove block range roots that have lower bound range strictly above the given slot number

fn optimize<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = StdResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Tell the store to perform optimization tasks if it needs them to improve its query performances